+++ common/ipc/zero_mq_handler.h
@@ -25,7 +25,7 @@ protected:
     // @unique_id代表一个ZeromqUnit的唯一id
     // @identify表示的是对方连接的标识符
     // @data是数据包
-    virtual void onData(uint32_t unique_id, std::string& identify, const std::string& data) {}
+    virtual void onData(uint64_t unique_id, std::string& identify, const std::string& data) {}
 
 protected:
     std::vector<ZeromqUnitPtr> m_unit_vec;
@@ -37,14 +37,14 @@ protected:
 class ZeromqRouterHandler : public ZeromqHandler {
 public:
     // 返回唯一id, 失败返回0
-    uint32_t listen(uint32_t id, const std::string& addr);
+    uint64_t listen(uint32_t id, const std::string& addr);
 
     // @unique_id代表一个ZeromqUnit的唯一id
     // @other_id对方的id
     // @data数据
-    int sendData(uint32_t unique_id, uint32_t other_id, const std::string& data);
+    int sendData(uint64_t unique_id, uint32_t other_id, const std::string& data);
 
-    int sendData(uint32_t unique_id, const std::string& identify, const std::string& data);
+    int sendData(uint64_t unique_id, const std::string& identify, const std::string& data);
 };
 
 /////////////////////////////////////////////////////
@@ -53,9 +53,9 @@ public:
 class ZeromqDealerHandler : public ZeromqHandler {
 public:
     // 返回唯一id, 失败返回0
-    uint32_t connect(uint32_t id, const std::string& addr);    
+    uint64_t connect(uint32_t id, const std::string& addr);    
 
     // @unique_id代表一个ZeromqUnit的唯一id
     // @data数据
-    int sendData(uint32_t unique_id, const std::string& data);
+    int sendData(uint64_t unique_id, const std::string& data);
 };
